Move core under src/ as lib/
