+++ src/page_01/problem_11.py
@@ -4,7 +4,7 @@ from mathtools import product
 from mathtools.combinatorics import recurrence
 from mathtools.iterable import take
 
-from core import gridutils
+from lib import gridutils
 
 GRID = """
 08 02 22 97 38 15 00 40 00 75 04 05 07 78 52 12 50 77 91 08
